requirements.txt
amqp==2.3.2
analytics-python==1.2.9
billiard==3.5.0.4
boto3==1.9.38
celery==4.2.1
celery-redbeat==0.11.1
django-atomic-signals==1.1.0
django-cors-headers==2.4.0
django-filter==1.1.0
git+https://github.com/djangonauts/django-hstore.git@aac755f587ff8ea95c1f43e7f3df8e339b848b6b
django-livefield==3.1.0
django-redis==4.9.0
django-storages==1.7.1
django-timezone-utils==0.13
Django==2.0.9
djangorestframework-hstore==1.3
djangorestframework==3.7.7
dnspython==1.15.0
docker==3.5.1
drf-extensions==0.4.0
gevent==1.3.7
greenlet==0.4.15
grpcio==1.16.0
hiredis==0.2.0
jinja2==2.10
jsonfield==2.0.2
jsonschema==2.6.0
kombu==4.2.1
lazy-object-proxy==1.3.1
munch==2.3.2
newrelic==4.4.1.104
pdfkit==0.6.1
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyopenssl==18.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-gcm==0.4
python-rapidjson==0.6.3
pytz==2018.7
PyYAML==3.13
py_zipkin==0.15.1
raven==6.9.0
redis==2.10.6
requests==2.20.0
requests-oauthlib==1.0.0
rest_condition==1.0.3
retrying==1.3.3
stripe==1.84.2
URLObject==2.4.3
uWSGI==2.0.17.1
zipstream==1.1.4
